[REFACTOR] Split into explanatory directories
In order to keep a good and understandable application hierarchy
throughout, beginning a split into groupings based on intent.
